Convert back slashes to forward slashes in import statement.
Usually an import of css is replaced with the css content, however sometimes it results in an @import statement.
Could not produce an example that generates an @import statement. In a current project of mine this happens to some css files imported from a jspm package. Not sure what triggers it (size/other references/location).
Therefor, no unit test was added